Task 10: extract shared utilities and hooks from components
- Move getByPath + orderBy from HeroTab.jsx to util.service.js (named exports)
- BestiaryTab.jsx now imports orderBy from util.service.js instead of its own copy
- Extract useFocusTrap from AppDialogs.jsx into src/hooks/useFocusTrap.js
- AppDialogs.jsx imports from the new hook module
